Close button on post-selection add site pages
